refactor: group into AccountingProvidersAccordion
